scripts/requirements.txt
requests==2.25.1
Pillow==8.3.2
